Add a CODEOWNERS file
As discussed on today's Docs call, adding a CODEOWNERS file per this email: https://lists.tungsten.io/g/docs/message/226

I probably should've sent a PR for this, but I don't want to do that until we've fully transitioned from Gerrit. So direct to master it is. Please don't do this going forward. We're only doing it now for the Gerrit -> GitHub transition.